Decoupled, refactored extractor, added persistence.
Feature extraction is now done using the Extractor() class,
and the entire feature set for a test/train directory is generated.
The directory is traversed depth-first and recursively.
The network weights are saved to a file, allowing to test without
re-training. The feature values are now saved in an XML file,
allowing to extract features for only newly added files.