Implemented autotrainer module for param sweep
New autotrainer.py allows to train the NN on a parameter
grid for a number of epochs. All progress is saved to a
log file, and all new network weights are written to an
xml file. After the sweep is complete, a confusion matrix
for the best set of parameters is created, showing the hit
accuracy for each audio class. The data set is split with
proportion 0.6.

The whatsound package is now actually a Python package.

Scripts inside core/ can still be executed as standalone,
but the autotrainer is exclusively a package module, and
must be run with the -m argument in the terminal.